Remove the need for scratch space when squaring
BN_sqr only has scratch space because it needs to compute a temporary
buffer with the a[i]^2 terms to add into the final result. But those
terms can be computed and added in a single pass.

This isn't expected to have any performance impact on assembly-enabled
builds. All those builds have bn_mul_mont optimizations, which means the
plain squaring operation is more-or-less unused. (This begs the question
why we have assembly optimizations for it, when it's only used in
conjunction with builds that barely use it, but ah well.) On NO_ASM
builds, the plain square operation is used more, but this impacts
linearly many terms out of an overall quadratic operation.

I was unable to measure a consistent difference with or without this
change. Really the benefit is that, by removing the dependency on
scratch space, we can remove the dependency on BN_CTX and can unify our
various Montgomery multiplication codepaths.

Bug: 42290433
Change-Id: I1527bd212529bbd4a1abedec22bb1dc3d7e12cbb
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/79307
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>